[REF] start using parcel bundler, use p5 in instance mode
This commit introduces the use of parcel as a bundler, in doing so, we
lose access to the global variables defined by p5, so sketch.js has been
modified to use p5's instance mode to accomodate.

This is in prepartion of splitting algorithms into their own files,
which will be made much easier using import syntax and automatic
bundling.